feat: move HistoryItem interface to common types
the historyitem interface is common among employee, laptops and any other
future module
